Move the zoom function tests to another class.
