+++ cameramodule/src/test/java/com/benine/ipcameracontrol/IpcameraFocusTest.java
@@ -17,6 +17,11 @@ import org.mockserver.verify.VerificationTimes;
 
 import com.benine.CameraConnectionException;
 
+/**
+ * Test class to test the IP Camera Focus function class.
+ * The mock server is used to simulate the camera.
+ * @author Bryan
+ */
 public class IpcameraFocusTest {
   
   @Rule
